+++ NEWS.md
@@ -1,5 +1,6 @@
 # medfate 4.8.3
 * Soil pool results (REW and psi) enabled
+* Areas changed by bars in PET_Precipitation plots
 
 # medfate 4.8.2
 * Bug correction detecting date vector objects (Rcpp issue)