+++ NEWS.md
@@ -3,6 +3,7 @@
 * Areas changed by bars in PET_Precipitation plots
 * New plotting function soil_conductivityCurvePlot()
 * Rhizosphere overlap is a parameter in function defaultControl()
+* Internal sureau structures rebuild
 
 # medfate 4.8.2
 * Bug correction detecting date vector objects (Rcpp issue)